requirements.txt
torch==2.1.0
transformers==4.31.0
gradio==3.35.1
accelerate==0.21.0
python-dotenv==1.0.0